restruct cli
