+++ src/cli/commands/checkpoint/list_checkpoints.rs
