--- src/cli/commands/manager/mod.rs
@@ -1,15 +1,0 @@
-// Copyright 2022-2023 Protocol Labs
-// SPDX-License-Identifier: MIT
-pub mod create;
-pub mod fund;
-pub mod join;
-pub mod kill;
-pub mod leave;
-pub mod list_checkpoints;
-pub mod list_subnets;
-pub mod net_addr;
-pub mod propagate;
-pub mod release;
-pub mod send_value;
-pub mod wallet;
-pub mod whitelist;
