+++ src/cli/commands/subnet/net_addr.rs
